Configuration Fixes
Removed settings directory, was trying to imitate hub setup but was more trouble
than it was worth, having settings and settings_heroku is just fine.

Had to change requirements.txt. For some reason my local env was not detecting the
integration settings package if installed via master.zip method. Will probably have
similar difficulties in the future with other packages, but leaving them as they are
and will cross that bridge when I come to it.