First FeinCMS Content Type Added
Created first custom content type MainSlider, and SliderImage upon which
it depends via foreign key relationship. This seemed the easiest way to
pair images and text/html to overlay on them, while also allowing greater
customization (any number of panes for the slider, reorder easily, etc.)

Page creation interface working correctly and recognizing my custom
content type. However, the image selector is currently empty. I need to
add to admin.py to create the tools necessary to create and manage that
content type.

Once that is complete and this is verified to render correctly, it's simply
a matter of duplicating these efforts for each other section of the unify
templates that we wish to use, and adding a WYSIWYG editor.