Footer Navigation
Navigation block for bottom of page, auto-populated from the page module
navigation tag.

Currently cannot find a way to override width of block to fill full window
like it is in the Hub, but it's a minor cosmetic thing we can patch later
if anyone really cares.

Cannot make a second row of links align to the left of the block, it is
forcing to center. Should not matter, shouldn't be more than 6 link trees
anyway, but just making note that this is a "would be nice to style properly"
item.