Issue #8: Header Styling
Removed title text as it will be redundant when the current generic logo
is replaced by the dedicated conference logo.

Greatly simplified styling with only two divs to worry about. Cannot seem
to get the menu to be at the bottom rather than the top, but I kind of
like it this way.