Issue #14 & #24
Created "Featured Image Link Pane" content type (best name I could think
of). Left layout styling for the layout templates, as it should be. Be
sure to always wrap the region where these go in a featured-blog class.

Also went through and cleaned up the existinc content types' verbose names
so the admin interface is a little easier to use.