Issue #17: Medium Image Content Type
Created medium image content type that allows for the creation of a two-
column layout with a rich text block opposite an image of the same size.

Did not create a content type for the whole layout piece, more modular if
we just create regions in the layout templates to drop these into, that way
we don't need to duplicate anything.

Also found a way to shift the gallery being rendered to center it better.