Home Page Template
First official template! home_page.html mimics exactly the layout of the
Unify page_home_14 template. Regions are created for two parallax boxes,
but currently that type is not functional.

Cleaned up local settings for caching. It was working a little too well,
forgot to include a local exception that points to a dummy.