Issues #33 & #32
Addressed clock styling and repositioned.

Reverted thumbnail tags for logo images. Cannot get sorl-thumbnail to cooperate with
S3, continues to return 403 or 404 errors (if AWS_QUERYSTRING_AUTH is off or on respectively).

Abandoning this attempt for now as it's a cosmetic issue. Recommending for now instructing
staff to scale images in photoshop before upload until someone else can find a solution.