Issue #96
Added minidetector package. Enabled middleware to add toggle variables
to the request context.

Modified nav menu to present differently based on the variable
"is_touch_device". If true, it's displayed as it has been, with the
menu opening on hover OR click, and the parent page link appearing
first in the dropdown.

If false, then the parent target is the page link, and it's removed
from the dropdown. The menu display toggle is switched to hover-only.