Header Bar Restyling
Finally got others on board to let me clean this up so it doesn't take up
half the screen.

Still not a fan of a new row for the dropdown menus, ideally I'd have that
just below the clock, but they have too many items in the menu on prod so
that it stretches underneath the block of text next to the logo.