Issue #109: Pricing Table Not Responsive
Change of plans. Can't find a way to make this scale, and it's not
best practice anyway. Instead, I make the standalone columns display
the registration type labels inline, as it was intended in the original
Unify template.

To get the correct labels, I simply changed the single price variables
where I'm packing the data into columns in the render method to instead
be dictionaries so I can associate them with level objects in the template
layer. Then I just swap {{ price }} with {{ price.price }} and nothing
changes, and I add a hidden span below for {{ price.level.level_name }}
labels.