Issue #110
Made number of columns in loop dynamic.

Added "strikethrough" field to deadline model to allow toggle to disable
a column. Checking the box causes the column to have struck out prices,
grays out the top and bottom boxes, and removes the CVENT registration
link and replaces it with "Deadline Has Passed" text.