Issue #112: Host Institution/Sponsor Content Type
Created content type for display of host institution and sponsor logos.

Enabled admin editing and created associated migrations. Enabled in templates.

With many different levels of sponsorship, it was easiest to add a level
selector field with defined choices. Then, in the render() method, choose
the template based on this level before rendering instead of trying to do
that logic within the template layer.