Registration Changes
Commented out registration button on base template now that online
registration has closed. Will investigate a way to make this controllable
via the admin panel in the future so I don't need to edit the code directly
every spring to turn it on and every fall to turn it off.

Added another conditional branch in the registration button fragment. Now
if a registration deadline's TITLE contains "Onsite", the button at the bottom
says "Register Onsite" and is a dummy link to target '#' rather than to cvent.